+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     
   - #### 🚀 Enhancements:
     - [x] Enhancement #866 to Improve performance of Input Info Analysis in 'Automatic Migration Feature' using an object from class FolderAnalyzer instead of performing read/write disk operations on LocalFolder class methods.
+    - [x] Enhancement on 'Album Pulled' / 'Album Pushed' dashboard messages during `Automatic Migration Feature`. Now Album messages are displayed in bright color to highlight it vs normal asset pull/push operations. 
 
   - #### 🐛 Bug fixes:
     - [x] Fixed Bug #884 in 'Google Takeout Processing' feature when flag `-gics, --google-ignore-check-structure` is detected causing that Output Folder is the same as Input Takeout Folder and deleting that at the end of the process.